moving map related function to src/components/mapbox/utils and initializer to src/components/mapbox/init
