+++ src/utils/dataset.tsx
@@ -1,16 +1,3 @@
-// Filter the type of reseller with the select in the top bar
-import { flyToStore } from './view';
-import { addPopup, removePopups } from '../components/Popup/Popup';
-import mapboxgl, { LngLatLike } from 'mapbox-gl';
-import { MapBoxListing, MountedMapsContextValue } from '../types';
-import { highlightListing } from '../components/Mapbox/utils';
-import * as MapboxGeocoder from '@mapbox/mapbox-gl-geocoder';
-import { RefObject } from 'react';
-import { SearchPopup } from '../components/Popup/PopupContent';
-import { SEARCH_RESULTS_SHOWN } from '../constants';
-import { useContext } from '@wordpress/element';
-import { MapboxContext } from '../components/Mapbox/MapboxContext';
-
 /**
  * The function prepares stores by assigning IDs to them and returning them as a geojson object.
  *
@@ -22,47 +9,6 @@ export const prepareStores = ( listings: any ) => ( {
 	type: 'geojson',
 	stores: listings,
 } );
-
-/**
- * The function filters stores based on selected tags and partnerships and populates the sidebar with
- * the filtered results.
- *
- * @param {Feature[]} stores - an array of store features, each of which is an object with properties such as
- *                           company, partnership, and location.
- * @param             terms
- * @return An array of strings that represent the keys of the partnerships in the input array.
- */
-export function filterStores( stores, terms ) {
-	const { searchResult, tag, partnership } = terms;
-
-	// Filter the stores based on the selected terms
-	const filteredStores = stores;
-
-	return filteredStores;
-}
-
-/**
- * Enables the listing feature on the map.
- *
- * @param {mapboxgl.Map}  map      - The map object.
- * @param {MapBoxListing} marker   - The listing marker object.
- * @param                 listings
- */
-export function enableListing(
-	map: mapboxgl.Map,
-	marker: MapBoxListing
-) {
-	console.log( 'Listing enabled', marker );
-
-	// 1. Fly to the point
-	flyToStore( map, marker );
-
-	// 2. Close all other popups and display popup for clicked store
-	addPopup( map, marker );
-
-	// 3. Highlight listing in sidebar (and remove highlight for all other listings)
-	highlightListing( marker );
-}
 
 /**
  * Given a JsonFeature array of objects, return the next ID to use
@@ -96,73 +42,3 @@ export const reorder = (
 
 	return result;
 };
-
-/**
- * Generates the marker properties for a geocoder marker.
- *
- * @param {MapBoxListing[]} stores      - The list of stores to generate the marker properties for.
- * @param {LngLatLike}      coordinates - The coordinates for the marker.
- * @return {Object} The marker properties.
- */
-export function geocoderMarkerProps(
-	stores: MapBoxListing[],
-	coordinates: LngLatLike
-): MapBoxListing {
-	return {
-		type: 'temp',
-		properties: {
-			name: 'geocoder',
-			icon: 'geocoder',
-			draggable: true,
-		},
-		geometry: {
-			type: 'Point',
-			coordinates,
-		},
-	};
-}
-
-/**
- * Displays the nearest store and updates the filtered listings.
- *
- * @param {MapboxGeocoder.Result}     location            - The location object.
- * @param {MapBoxListing[]}           sortedNearestStores - The sorted list of nearest stores.
- * @param {MapBoxListing[]}           filteredListings    - The filtered list of listings.
- * @param {RefObject<HTMLDivElement>} mapRef              - The reference to the map container.
- * @param {mapboxgl.Map}              map                 - The map object.
- */
-export function showNearestStore(
-	location: MapboxGeocoder.Result,
-	sortedNearestStores: MapBoxListing[],
-	mapRef: RefObject< HTMLDivElement >,
-	map: mapboxgl.Map
-): MapBoxListing[] {
-	const newFilteredListings: MapBoxListing[] = [
-		...sortedNearestStores.slice( 0, SEARCH_RESULTS_SHOWN ),
-	];
-
-	removePopups( mapRef as RefObject< HTMLDivElement > );
-
-	/* Open a popup for the closest store. */
-	addPopup(
-		map,
-		location,
-		<SearchPopup
-			icon={ 'geocoder' }
-			name={ location.text }
-			category={ location.properties?.category }
-			maki={ location.properties?.maki }
-			draggable={ true }
-		/>
-	);
-
-	/* Open a popup for the closest store. */
-	addPopup( map, newFilteredListings[ 0 ] );
-
-	/** Highlight the listing for the closest store. */
-	mapRef?.current
-		?.querySelector( '#marker-' + sortedNearestStores[ 1 ].id )
-		?.classList.add( 'active' );
-
-	return newFilteredListings;
-}
